fix: refactored service to mail provider and tests adjusted accordingly
